v2/requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
transformers
huggingface_hub
ctransformers
evaluate
bert_score
pdfminer.six==20221105
accelerate